DAO-212 Move pages/components/policies/pagination -> src/components/pagination
