+++ src/components/pagination/pagination.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/variables.module.scss';
+@import '../../styles/variables.module.scss';
 
 .nav {
   display: flex;
